controllers: allow replica count to update without revision change
We used to restart every pod in a statefulset (hbasemaster and/or
regionserver) when the replica count changed, instead of simply
allowing statefulset to expand. This is especially unnecessary when
there is no configuration change and the only change was replica
count.

In this change we do not use new replica consider as part of the
revision SHA calculation, but rather reuse the existing replica
count.

Test suite had to be updated to not use any assertions but rather
rely on Eventually() for state verification. This is because
assertions failed randomly, and that was based on timing - if
assertion happened to run when reconciler had not finished updating,
that assertion test would fail. This change removes the dependencies on
timing based issues.