controllers: improve logging message when regions are drained (#20)
During recent HBase upgrade we ran into an incident. Folks that had
to respond to the incident were quite confused by the logged message
when regions are being drained from a regionserver about to be
deleted/updated, thinking there was something wrong with how the
operator was draining regions. Improve the logged message to help
reduce confusion.